starting main view refactoring
